requirements.txt
diff-match-patch==20121119
Django==2.0.1
django-disqus==0.5
django-import-export==0.7.0
django-nocaptcha-recaptcha==0.0.20
django-suit==0.2.25
django-wysiwyg-redactor==0.5.1
et-xmlfile==1.0.1
freeze==1.0.10
jdcal==1.3
mock==2.0.0
odfpy==1.3.6
openpyxl==2.4.9
pbr==1.10.0
Pillow==5.0.0
PyMySQL==0.8.0
pytz==2016.6.1
PyYAML==3.12
six==1.10.0
tablib==0.12.1
unicodecsv==0.14.1
xlrd==1.1.0
xlwt==1.3.0
